+++ src/main/java/de/muenchen/mobidam/scheduler/SchedulerContext.java
@@ -1,0 +1,54 @@
+/*
+ * The MIT License
+ * Copyright © 2024 Landeshauptstadt München | it@M
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+package de.muenchen.mobidam.scheduler;
+
+import de.muenchen.mobidam.Constants;
+import de.muenchen.mobidam.config.Interfaces;
+import org.quartz.JobDetail;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.quartz.CronTriggerFactoryBean;
+import org.springframework.scheduling.quartz.JobDetailFactoryBean;
+
+@Configuration
+public class SchedulerContext {
+
+    @Bean
+    public JobDetailFactoryBean parkAndRideDetail() {
+        var jobDetailFactory = new JobDetailFactoryBean();
+        jobDetailFactory.setJobClass(ParkAndRideStaticJob.class);
+        jobDetailFactory.setDescription("Mobilithek Info Park And Ride Static Job");
+        jobDetailFactory.setDurability(true);
+        return jobDetailFactory;
+    }
+
+    @Bean
+    public CronTriggerFactoryBean cronParkAndRide(JobDetail job, Interfaces properties) {
+        var cron = new CronTriggerFactoryBean();
+        cron.setJobDetail(job);
+        cron.setGroup("Park And Ride Static Job");
+        cron.setCronExpression(properties.getInterfaces().get(Constants.PARK_RIDE_STATIC_DATA).getCronExpression());
+        return cron;
+    }
+
+}