+++ RELEASENOTES.md
@@ -1,5 +1,9 @@
 # Release-Notes
 
+## Sprint 11 + 12 (08.05.2024 - 18.06.2024)
+### Hinzugefügt
+- Anbindung Mobilithek, S3 und SST-Management erstellt.
+
 ## Sprint 10 (16.04.2024 - 07.05.2024)
 ### Hinzugefügt
 - Workflows erstellt